+++ interfaces/kits/js/napi/src/is_last_work_time_out.cpp
@@ -40,7 +40,7 @@ struct AsyncCallbackIsLastWorkTimeOut {
 
 napi_value ParseParameters(const napi_env &env, const napi_callback_info &info, IsLastWorkTimeOutParamsInfo &params)
 {
-    size_t argc = IS_LAST_WORK_TIME_OUT_MAX_PARAMS;
+    int32_t argc = IS_LAST_WORK_TIME_OUT_MAX_PARAMS;
     napi_value argv[IS_LAST_WORK_TIME_OUT_MAX_PARAMS] = {nullptr};
     NAPI_CALL(env, napi_get_cb_info(env, info, &argc, argv, NULL, NULL));
     bool paramsCheck = (argc == IS_LAST_WORK_TIME_OUT_MAX_PARAMS) || (argc == IS_LAST_WORK_TIME_OUT_MIN_PARAMS);
